+++ src/app/services/api/index.ts
@@ -1,12 +1,12 @@
-import { angularPath, awsPath, coursesData, coursesPath, jsPath } from '../data/courses-data';
-import { type DataMap } from '../data/courses-data.types';
+import { courses, coursesPath, jsPath, angularPath, awsDevPath } from '../data';
+import { type DataMap } from '../data';
 
 const dataProviders: DataMap = {
-  courses: coursesData,
+  courses: courses,
   coursesPath: coursesPath,
   javascriptPath: jsPath,
   angularPath: angularPath,
-  awsDevPath: awsPath
+  awsDevPath: awsDevPath
 };
 
 export const fetchDataByName = (dataName: keyof DataMap) => {
