+++ src/app/services/data/angular-path.data.ts
@@ -1,0 +1,60 @@
+import { AngularAwsPath } from './courses-data.types';
+
+export const angularPath: AngularAwsPath[] = [
+  {
+    id: 1,
+    title: 'Week #1',
+    actions: [
+      'Module "Angular Intro. TypeScript."',
+      'Module "Angular. Components"',
+      'Module "Angular. Directives & Pipes"'
+    ]
+  },
+  {
+    id: 2,
+    title: 'Week #2',
+    actions: [
+      'Module "Angular Intro. Task Review."',
+      'Angular. Modules & Services, Dependency Injection',
+      'Module "Angular. Directives & Pipes"',
+      'Module "Angular. Routing"',
+      'Workshop'
+    ]
+  },
+  {
+    id: 3,
+    title: 'Week #3',
+    actions: [
+      '"Angular. Components, Directives, Pipes" task review',
+      'Module "RxJS & Observables"',
+      'Module "Angular. HTTP"',
+      'Module "Angular. Forms"',
+      'Workshop'
+    ]
+  },
+  {
+    id: 4,
+    title: 'Week #4',
+    actions: [
+      '"Angular. Modules, Services, Routing" task review',
+      'Module "Angular. Redux & NgRx"',
+      'Module "Angular. Unit Test"',
+      'Workshop'
+    ]
+  },
+  {
+    id: 5,
+    title: 'Week #5-8',
+    actions: [
+      '"Angular. RxJS & HTTPClient & NgRx & Forms", task review',
+      'Final Angular test',
+      'Workshop',
+      '"Project Management Application" final task'
+    ]
+  },
+  {
+    id: 5,
+    title: 'Week #9',
+    actions: ['Cross-checking the "Project management application" final task']
+  }
+];
