+++ src/app/services/data/index.ts
@@ -1,0 +1,6 @@
+export { angularPath } from './angular-path.data';
+export { awsDevPath } from './aws-dev-path.data';
+export { jsPath } from './js-path.data';
+export { coursesPath } from './courses-path.data';
+export { courses } from './courses.data';
+export { type DataMap } from './courses-data.types';
